skill factory
now is in skillfactory script is an enum of skills and list of skills, this enum must be same as type of action in list. now can be easily added skill runtime, and in inspector, other things like attack, projectile speed, projectile prefab, action cost...etc is based on data in characterstats. idea is create skills based on sciptable object, so in future there will be only scriptable object in characterstats(maybe new script Character?)